move user and portable defs into db
